requirements.txt
flask==2.0.1
openai-whisper
torch
opencc-python-reimplemented 